chore(deps): gitlab.com/gitlab-org/api/client-go from 0.127.0 to 1.46.0 (#753)
* Bump gitlab.com/gitlab-org/api/client-go from 0.127.0 to 1.46.0

Bumps [gitlab.com/gitlab-org/api/client-go](https://gitlab.com/gitlab-org/api/client-go) from 0.127.0 to 1.46.0.
- [Release notes](https://gitlab.com/gitlab-org/api/client-go/tags)
- [Changelog](https://gitlab.com/gitlab-org/api/client-go/blob/main/CHANGELOG.md)
- [Commits](https://gitlab.com/gitlab-org/api/client-go/compare/v0.127.0...v1.46.0)

---
updated-dependencies:
- dependency-name: gitlab.com/gitlab-org/api/client-go
  dependency-version: 1.46.0
  dependency-type: direct:production
  update-type: version-update:semver-major
...


* fix: MR IID type to int64

* fix: MR IID type to int64 in tests

---------

+++ internal/gitlab/gitlab.go
@@ -129,7 +129,7 @@ func (c *GitlabConfig) MergeRequestsForCommit(commit string) ([]*gitlab.BasicMer
 }
 
 // GetMergeRequestApprovers returns a list of users (name and username) who approved an MR
-func (c *GitlabConfig) GetMergeRequestApprovers(mrIID, version int) ([]any, error) {
+func (c *GitlabConfig) GetMergeRequestApprovers(mrIID, version int64) ([]any, error) {
 	var approvers []any
 	client, err := c.NewGitlabClientFromToken()
 	if err != nil {
